D files now generated through a callback url
